Modified unique options handling in argparser and written most of option handling functions/macros in client and other minor modifies

+++ src/include/dir_utils.h
@@ -4,7 +4,7 @@
 #include <defines.h>
 #include <linkedlist.h> 
 
-int dirscan(const char nomedir[], int n, llist_t** filelist);
+int dirscan(const char nomedir[], long n, llist_t** filelist);
 
 int loadFile(const char* pathname, void** buf, size_t* size);
